+++ Modules/Intent.Modules.AspNetCore.Swashbuckle/release-notes.md
@@ -1,3 +1,4 @@
 ### Version 4.0.0
 
 - New: Moved configuration from `appsettings.json` to code configuration.
+- New: Upgraded Templates to use new Builder Pattern paradigm.